feat(debugger): allow examples to be predefined
