+++ .changeset/eleven-foxes-tickle.md
@@ -1,0 +1,7 @@
+---
+"template-next-starter-with-examples": patch
+"@frames.js/debugger": patch
+"create-frames": patch
+---
+
+feat(debugger): allow examples to be predefined
