+++ templates/next-starter-with-examples/scripts/dev-script.js
@@ -1,6 +1,16 @@
 import "dotenv/config";
 import { spawn } from "node:child_process";
+import { readdir } from "node:fs/promises";
 import isPortReachable from "is-port-reachable";
+import { fileURLToPath } from "node:url";
+import { resolve, dirname } from "node:path";
+
+function snakeCaseToTitleCase(snakeCase) {
+  return snakeCase
+    .split("-")
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+    .join(" ");
+}
 
 async function getOpenPort(port) {
   const isReachable = await isPortReachable(port, { host: "localhost" });
@@ -10,6 +20,28 @@ async function getOpenPort(port) {
   }
 
   return port;
+}
+
+/**
+ * @param {number} port
+ * @returns {Promise<{ title: string; url: string; }[]>}
+ */
+async function getExamplesFromDirectory(port) {
+  const currentDirectory = dirname(fileURLToPath(import.meta.url));
+  const examplesDirectory = resolve(currentDirectory, "../app/examples");
+
+  const foundDirectoriesAndFilesInExamplesDirectory = await readdir(
+    examplesDirectory,
+    { withFileTypes: true }
+  );
+  const exampleDirectories = foundDirectoriesAndFilesInExamplesDirectory
+    .filter((dirent) => dirent.isDirectory())
+    .map((dirent) => dirent.name);
+
+  return exampleDirectories.map((example) => ({
+    title: snakeCaseToTitleCase(example),
+    url: `http://localhost:${port}/examples/${example}`,
+  }));
 }
 
 const nextPort = await getOpenPort(3000);
@@ -22,6 +54,10 @@ process.env.DEBUGGER_HUB_HTTP_URL = `http://localhost:${debuggerPort}/hub`;
 
 if (!process.env.FJS_MONOREPO) {
   const url = `http://localhost:${nextPort}`;
+
+  const examples = await getExamplesFromDirectory(nextPort);
+
+  process.env.DEBUGGER_EXAMPLES_JSON = JSON.stringify(examples);
 
   command = "concurrently";
   args = [
